# Improved Sliding algorithm for ExponentialSpringForce