… Simbody/src/ExponentialSpringParameters.cpp
… Simbody/tests/TestExponentialSpring.cpp
… Simbody/tests/adhoc/ExponentialSpringsComparison.cpp