@sherm1 

This PR includes:
1. A new algorithm for handling the Sliding State.
2. A modification to the API that I believe will be more intuitive for the user. Specifically, the Friction Spring Zero is now consistently referred to as the Elastic Anchor Point (a term first suggested by @sherm1).
3. Improved initial conditions and comments in the ```Test_Adhoc - ExponentialSpringsComparison``` program.

The chief improvement in this PR is a new algorithm for handling the ```Sliding``` State of the ```ExponentialSpringForce``` contact model ( 1.  in the list above).  The ```Sliding``` State (now ```K```) ```(0.0 ≤ K ≤ 1.0)``` characterizes the extent to which either static or kinetic conditions are present for a contact point. The contact point is referred to in the code as the Elastic Anchor Point (or Friction Spring Zero).  ```K = 0.0``` indicates static conditions (i.e., the anchor point is fixed in place).  ```K = 1.0``` indicates kinetic conditions (i.e., the anchor point is sliding). A value between 0.0 and 1.0 indicates a transition between the two extremes.

The new algorithm has allowed a fair amount of code to be removed, is about 10% faster, and I believe is more accurate.

In the previous version of ```ExponentialSpringForce```, ```Sliding``` was handled as a continuous state (a "Z").  It was modeled using a simple differential equation for a rising/decaying exponential.  Notably, the transition between a value of 0.0 and 1.0, once triggered, was entirely decoupled from the dynamics of the body.

In the new algorithm, Sliding is handled as an auto update discrete variable (```K```).  It is calculated using a continuous function of the ratio of the speed of the elastic anchor point (```ṗ₀```) to the settle velocity (```vSettle```). [```vSettle``` is a parameter that represents the speed at which a body station settles into a static state.] Specifically,

```
        ṗ₀ = |Δp₀| / Δt
        K = SimTK::stepUp( SimTK::clamp(0.0, ṗ₀/vSettle, 1.0) )
```

where ```Δp₀``` is the change in ```p₀``` and ```Δt``` is the change in time since the last successful integration step. When ```ṗ₀ ≥ vSettle```, ```K = 1.0```, and as ```ṗ₀ → 0.0```, ```K → 0.0```.

Presented below is a file-by-file description of the changes that were made.

**ExponentialSpringParameters.h**
- Parameters for the Sliding Time Constant (```tau```) and Settle Acceleration (```aSettle```), as well as their related get and set methods were removed. These parameters are not used by the new sliding algorithm.
- In user documentation, the "friction spring zero" is now referred to as the "elastic anchor point".

**ExponentialSpringParameters.cpp**
- Accessor implementations for tau and ```aSettle``` have been removed.
- ```tau``` and ```aSettle``` member variables are no longer referenced in the constructor and the ```opperator==``` method.

**ExponentialSpringForce.h**
- Doxygen comments for class ```ExponentialSpringForce``` have been updated to explain the new algorithm for the ```Sliding``` State and now use the phrase "elastic anchor point" in preference to "friction spring zero".  In addition, a few tweaks were made to generally improve the clarity of the comments.
- ```setSliding()``` was removed.
- ```resetFrictionSpringZero()``` was changed to ```resetAnchorPoint()```.
- ```getFrictionSpringZeroPosition()``` was changed to ```getAnchorPointPosition()```.

**ExponentialSpringForce.cpp**
- Anything related to the ```SlidingAction``` state was removed.
- The ```Sliding``` State, formerly ```Sliding``` and now ```K```, was changed from a continuous state to an auto update discrete state.
- ```ExponentialSpringForceImpl::realizeAcceleration()``` was removed.
- ```SprZero``` and ```FrictionSpringZero``` were changed to ```AnchorPoint``` in method and variable names.
- The new algorithm for computing the ```Sliding``` State (```K```) can be found near the end of ```ExponentialSpringForceImpl::realizeDynamics()```.
- ```ExponentialSpringForce::setSliding()``` was removed.
- ```ExponentialSpringForce::getSliding()``` was moved to the end of the file next to ```ExponentialSpringForce::getAnchorPointPosition()``` so that accessors for the auto update discrete states would be next to each other.
- A number of comments were tweaked to improve clarity.

**ExponentialSpringsComparison.cpp**
- Initial conditions for the simulations were altered to produce simulations that give more informative comparisons between ```ExponentialSpringForce``` and ```CompliantContactSubsystem```.
- A variable, ```vTransition```, was added to ```main()``` to make this variable's value more conspicuous to a potential user. Comments emphasizing and explaining the importance of the variable were also added.
- ```resetSpringZero()``` was changed to ```resetAnchorPoint()```.
- Maximum integration step size was changed from 0.05 to 0.03 sec.
- The final time of the simulation was changed from 10.0 to 12.0 sec.

**TestExponentialSpring.cpp**
- ```resetSpringZero()``` was changed to ```resetAnchorPoint()```
- ```getFrictionSpringZeroPosition()``` was changed to ```getAnchorPointPosition()```
- The value of ```Sliding``` now reliably stays between 0.0 and 1.0 and so the ```SimTK_TEST``` does not need to allow for overshoot.
- ```SimTK_TEST``` calls relating to the Sliding Time Constant and Settle Acceleration were removed because those parameters no longer exist.

<!-- Reviewable:start -->
- - -
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/simbody/simbody/746)
<!-- Reviewable:end -->